requirements.txt
pandas
numpy
scikit-learn
python-dotenv
jupyter
google-genai
pydantic